# Use worker for top progress bar