… src/ElectronBackend/input/OpossumOutputFileSchema.json
… src/Frontend/Components/ProgressBar/__tests__/TopProgressBar.test.tsx
… src/Frontend/state/actions/resource-actions/__tests__/all-views-simple-actions.test.ts
… src/Frontend/state/actions/resource-actions/__tests__/load-actions.test.ts
… src/Frontend/state/actions/resource-actions/__tests__/save-actions.test.ts
… src/Frontend/state/helpers/__tests__/progress-bar-data-helpers.test.ts
… src/Frontend/state/reducers/resource-reducer.ts
… src/Frontend/web-workers/folder-progress-bar-worker.ts