@benedikt-richter 2022-11-29
The file`package-lock.json` should not be checked in