@MarkusObendrauf 2022-11-30
> How do we handle sign offs? There are a couple of them missing, but I guess it is fine if the merge commit is signed?

I will squash it into a single commit before merging, that will be signed